refactor: separating action graph and adding activation graph
