+++ src/main/kotlin/at/ac/uibk/dps/cirrina/execution/graph/ActivationGraph.kt
@@ -1,0 +1,3 @@
+package at.ac.uibk.dps.cirrina.execution.graph
+
+class ActivationGraph {}
