+++ src/main/kotlin/at/ac/uibk/dps/cirrina/spec/Transition.kt
@@ -1,8 +1,8 @@
 package at.ac.uibk.dps.cirrina.spec
 
 import at.ac.uibk.dps.cirrina.csm.Csml.TransitionDescription
+import at.ac.uibk.dps.cirrina.execution.graph.ActionGraph
 import at.ac.uibk.dps.cirrina.execution.`object`.Action
-import at.ac.uibk.dps.cirrina.execution.`object`.ActionGraph
 import at.ac.uibk.dps.cirrina.execution.`object`.Extent
 import at.ac.uibk.dps.cirrina.execution.`object`.Guard
 import org.jgrapht.graph.DefaultEdge
